# Replace some usages of fmt.Sprintf with more efficient string concatenation