--- a/operator/pkg/model/ingestion/ingress.go
+++ b/operator/pkg/model/ingestion/ingress.go
@@ -4,7 +4,6 @@
 package ingestion
 
 import (
-	"fmt"
 	"sort"
 
 	"github.com/cilium/cilium/operator/pkg/ingress/annotations"
@@ -87,7 +86,7 @@ func Ingress(ing slim_networkingv1.Ingress, defaultSecretNamespace, defaultSecre
 		l.Port = 80
 		l.Sources = model.AddSource(l.Sources, sourceResource)
 		if !ok {
-			l.Name = fmt.Sprintf("ing-%s-%s-%s", ing.Name, ing.Namespace, host)
+			l.Name = "ing-" + ing.Name + "-" + ing.Namespace + "-" + host
 		}
 
 		l.Hostname = host