--- a/pkg/kvstore/allocator/allocator.go
+++ b/pkg/kvstore/allocator/allocator.go
@@ -393,7 +393,7 @@ func (k *kvstoreBackend) RunLocksGC(ctx context.Context, staleKeysPrevRound map[
 	for key, v := range allocated {
 		scopedLog := log.WithFields(logrus.Fields{
 			fieldKey:     key,
-			fieldLeaseID: fmt.Sprintf("%x", v.LeaseID),
+			fieldLeaseID: strconv.FormatUint(uint64(v.LeaseID), 16),
 		})
 		// Only delete if this key was previously marked as to be deleted
 		if modRev, ok := staleKeysPrevRound[key]; ok &&
@@ -443,7 +443,7 @@ func (k *kvstoreBackend) RunGC(
 
 	min := uint64(minID)
 	max := uint64(maxID)
-	reasonOutOfRange := fmt.Sprintf("out of local cluster identity range [%d,%d]", min, max)
+	reasonOutOfRange := "out of local cluster identity range [" + strconv.FormatUint(min, 10) + "," + strconv.FormatUint(max, 10) + "]"
 
 	// iterate over /id/
 	for key, v := range allocated {