As mentioned in https://github.com/cilium/cilium/issues/19571 `fmt.Sprintf()` is well-known for not being opimal in terms of CPU and memory allocations. This PR replaces some usages of `fmt.Sprintf` on possible hot code paths with more efficient string concatenation.

* [pkg/policy] Benchmark ProxyID
* [pkg/policy] ProxyID uses string concat instead of Sprintf
* [pkg/policy] L4PolicyMap uses string concatenation
* [pkg/policy] map state uses string concatenation
* [pkg/policy] rulePortsCoverSearchContext uses string concatenation
* [pkg/policy] NewVisibilityPolicy uses string concatenation
* [pkg/policy] groups use string concatenation
* [pkg/loadbalancer] Benchmark L3n4Addr String and StringWithProtocol
* [pkg/loadbalancer] L3n4Addr String and StringWithProtocol use string concatenation
* [pkg/kvstore] makeSessionName and RunLocksGC use string concatenation
* [pkg/endpoint] Use string concatenation
* [operator/pkg/model] Ingress listener name uses string concatenation

Related: https://github.com/cilium/cilium/issues/19571



- [ ] For first time contributors, read [Submitting a pull request](https://docs.cilium.io/en/stable/contributing/development/contributing_guide/#submitting-a-pull-request)
- [ ] All code is covered by unit and/or runtime tests where feasible.
- [ ] All commits contain a well written commit description including a title,
      description and a `Fixes: #XXX` line if the commit addresses a particular
      GitHub issue.
- [ ] If your commit description contains a `Fixes: <commit-id>` tag, then
      please add the commit author[s] as reviewer[s] to this issue.
- [ ] All commits are signed off. See the section [Developer’s Certificate of Origin](https://docs.cilium.io/en/stable/contributing/development/contributing_guide/#dev-coo)
- [ ] Provide a title or release-note blurb suitable for the release notes.
- [ ] Are you a user of Cilium? Please add yourself to the [Users doc](https://github.com/cilium/cilium/blob/main/USERS.md)
- [ ] Thanks for contributing!


```release-note
<!-- Enter the release note text here if needed or remove this section! -->
```