@schlosna 2023-08-15
`benchstat` output, old is (main) ec9cb2393601615c5db211ced1e558e4cce6790b + 2 benchmark commits (https://github.com/cilium/cilium/pull/27518/commits/8d188a5e6ce956f98f61fdae8647ae700ad69d8a + https://github.com/cilium/cilium/pull/27518/commits/5ddd7a41662de8947bf9b861303fd3e6e98f84e3 )

```
goos: darwin
goarch: arm64
pkg: github.com/cilium/cilium/pkg/loadbalancer
                                        │ old-loadbalancer.txt │         new-loadbalancer.txt         │
                                        │        sec/op        │    sec/op     vs base                │
L3n4Addr_Hash_IPv4-10                             21.63n ±  1%   22.10n ± 13%   +2.17% (p=0.001 n=10)
L3n4Addr_Hash_IPv6_Short-10                       21.63n ±  1%   21.95n ±  3%   +1.48% (p=0.007 n=10)
L3n4Addr_Hash_IPv6_Long-10                        21.81n ±  2%   22.00n ±  3%        ~ (p=0.210 n=10)
L3n4Addr_Hash_IPv6_Max-10                         22.03n ± 11%   21.97n ± 10%        ~ (p=0.986 n=10)
L3n4Addr_String_IPv4-10                          150.65n ±  1%   74.55n ± 13%  -50.52% (p=0.000 n=10)
L3n4Addr_String_IPv6_Max-10                       183.3n ± 15%   121.7n ± 26%  -33.62% (p=0.000 n=10)
L3n4Addr_StringWithProtocol_IPv4-10              186.65n ± 11%   79.98n ±  5%  -57.15% (p=0.000 n=10)
L3n4Addr_StringWithProtocol_IPv6_Max-10           213.3n ±  4%   125.4n ± 14%  -41.23% (p=0.000 n=10)
geomean                                           62.97n         46.36n        -26.38%

                                        │ old-loadbalancer.txt │         new-loadbalancer.txt         │
                                        │         B/op         │    B/op     vs base                  │
L3n4Addr_Hash_IPv4-10                               24.00 ± 0%   24.00 ± 0%        ~ (p=1.000 n=10) ¹
L3n4Addr_Hash_IPv6_Short-10                         24.00 ± 0%   24.00 ± 0%        ~ (p=1.000 n=10) ¹
L3n4Addr_Hash_IPv6_Long-10                          24.00 ± 0%   24.00 ± 0%        ~ (p=1.000 n=10) ¹
L3n4Addr_Hash_IPv6_Max-10                           24.00 ± 0%   24.00 ± 0%        ~ (p=1.000 n=10) ¹
L3n4Addr_String_IPv4-10                             74.00 ± 0%   44.00 ± 0%  -40.54% (p=0.000 n=10)
L3n4Addr_String_IPv6_Max-10                         114.0 ± 0%   101.0 ± 0%  -11.40% (p=0.000 n=10)
L3n4Addr_StringWithProtocol_IPv4-10                 98.00 ± 0%   52.00 ± 0%  -46.94% (p=0.000 n=10)
L3n4Addr_StringWithProtocol_IPv6_Max-10             146.0 ± 0%   117.0 ± 0%  -19.86% (p=0.000 n=10)
geomean                                             50.16        41.60       -17.06%
¹ all samples are equal

                                        │ old-loadbalancer.txt │         new-loadbalancer.txt         │
                                        │      allocs/op       │ allocs/op   vs base                  │
L3n4Addr_Hash_IPv4-10                               1.000 ± 0%   1.000 ± 0%        ~ (p=1.000 n=10) ¹
L3n4Addr_Hash_IPv6_Short-10                         1.000 ± 0%   1.000 ± 0%        ~ (p=1.000 n=10) ¹
L3n4Addr_Hash_IPv6_Long-10                          1.000 ± 0%   1.000 ± 0%        ~ (p=1.000 n=10) ¹
L3n4Addr_Hash_IPv6_Max-10                           1.000 ± 0%   1.000 ± 0%        ~ (p=1.000 n=10) ¹
L3n4Addr_String_IPv4-10                             5.000 ± 0%   3.000 ± 0%  -40.00% (p=0.000 n=10)
L3n4Addr_String_IPv6_Max-10                         4.000 ± 0%   3.000 ± 0%  -25.00% (p=0.000 n=10)
L3n4Addr_StringWithProtocol_IPv4-10                 6.000 ± 0%   3.000 ± 0%  -50.00% (p=0.000 n=10)
L3n4Addr_StringWithProtocol_IPv6_Max-10             5.000 ± 0%   3.000 ± 0%  -40.00% (p=0.000 n=10)
geomean                                             2.225        1.732       -22.14%
¹ all samples are equal
```