@joestringer 2023-08-25
It looks like this subtly changed LB frontend string formatting which broke a bunch of golden test cases: https://app.travis-ci.com/github/cilium/cilium/builds/265504327 .

I've also kicked off the full CI to get feedback from the rest of the testsuite. Those results should come in over the next 30-60 minutes.