@schlosna 2023-08-25
> It looks like this subtly changed LB frontend string formatting which broke a bunch of golden test cases: https://app.travis-ci.com/github/cilium/cilium/builds/265504327 .

Sorry about those failures. I'm adding some unit tests for LB `String()` and `StringWithProtocol()` and fixing these breaks.