@schlosna 2023-08-28
@christarazi I see several `Conformance` tests failed. I'm not very familiar with cilium CI yet, so not sure how to interpret these. From looking through the logs it appears to me that there are some connection failures. What is the the best way to debug and iterate on these types of CI failures?